Initializing the basis of the application
