+++ app/Makefile
@@ -1,0 +1,6 @@
+.PHONY: setup-autobs
+setup-autobs:
+	cd autobs && make all
+
+.PHONY: setup
+setup: setup-autobs
