+++ app/autobs/.gitignore
@@ -1,0 +1,2 @@
+./autobs
+./autobs_unix
