+++ app/autobs/Makefile
@@ -1,0 +1,28 @@
+# Go parameters
+GOCMD=go
+GOBUILD=$(GOCMD) build
+GOCLEAN=$(GOCMD) clean
+GOTEST=$(GOCMD) test
+GOGET=$(GOCMD) get
+BINARY_NAME=autobs
+BINARY_UNIX=$(BINARY_NAME)_unix
+
+all: test build
+build: 
+	$(GOBUILD) -o $(BINARY_NAME) -v ./cmd/autobs
+test: 
+	$(GOTEST) -v ./...
+clean: 
+	$(GOCLEAN)
+	rm -f $(BINARY_NAME)
+	rm -f $(BINARY_UNIX)
+run:
+	$(GOBUILD) -o $(BINARY_NAME) -v ./cmd/autobs
+	./$(BINARY_NAME)
+deps:
+	$(GOGET) ./...
+# Cross compilation
+build-linux:
+	CGO_ENABLED=0 GOOS=linux GOARCH=amd64 $(GOBUILD) -o $(BINARY_UNIX) -v ./cmd/autobs
+docker-build:
+	docker build -t $(BINARY_NAME) .
