+++ app/autobs/Readme.md
@@ -1,0 +1,3 @@
+# Automated Observer
+
+This project implements a set of services for running an automated observer, or a system to scan, parse, and filter external information into a more concise form.
