+++ app/autobs/go.mod
@@ -1,4 +1,4 @@
-module valesordev/autobs
+module github.com/valesordev/autobs
 
 go 1.22.2
 
@@ -6,6 +6,9 @@ require (
 	github.com/felixge/httpsnoop v1.0.4 // indirect
 	github.com/go-logr/logr v1.4.1 // indirect
 	github.com/go-logr/stdr v1.2.2 // indirect
+	github.com/inconshreveable/mousetrap v1.1.0 // indirect
+	github.com/spf13/cobra v1.8.0 // indirect
+	github.com/spf13/pflag v1.0.5 // indirect
 	go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp v0.51.0 // indirect
 	go.opentelemetry.io/otel v1.26.0 // indirect
 	go.opentelemetry.io/otel/exporters/stdout/stdoutmetric v1.26.0 // indirect
@@ -16,3 +19,7 @@ require (
 	go.opentelemetry.io/otel/trace v1.26.0 // indirect
 	golang.org/x/sys v0.19.0 // indirect
 )
+
+require internal/otel v1.0.0
+
+replace internal/otel => ./internal/otel
