+++ app/autobs/internal/otel/go.mod
@@ -1,0 +1,3 @@
+module otel
+
+go 1.22.2
