+++ infra/Makefile
@@ -1,0 +1,30 @@
+.PHONY: setup-k8s
+setup-k8s:
+	cd k8s && make setup
+
+.PHONY: setup-metrics
+setup-metrics:
+	cd infra/metrics && make setup
+
+.PHONY: setup-eventstore
+setup-eventstore:
+	cd infra/eventstore && make setup
+
+.PHONY: setup-search
+setup-search:
+	cd infra/search && make setup
+
+.PHONY: setup-alloy
+setup-alloy:
+	cd infra/alloy && make setup
+
+.PHONY: setup-cache
+setup-cache:
+	cd infra/cache && make setup
+
+.PHONY: setup-tabular
+setup-tabular:
+	cd infra/tabular && make setup
+
+.PHONY: setup
+setup: setup-k8s setup-metrics setup-eventstore setup-search setup-alloy setup-cache setup-tabular
